+++ docs/src/changelog.md
@@ -6,7 +6,8 @@ Release notes:
 
 ### Changed
 
--
+- `get_amsr()` handles all 4 data types (`daily`, `3day`, `weekly` and `monthly`)
+- `plot_amsr()` handles all 4 data types
 
 - BREAKING